working on reviews and meta

+++ src/main/kotlin/coraythan/keyswap/decks/models/DeckSearchResult.kt
@@ -45,7 +45,7 @@ data class DeckSearchResult(
         val synergyRating: Int = 0,
         val antisynergyRating: Int = 0,
         val adaptiveScore: Int = 0,
-        val metaScores: Map<String, Double> = mapOf(),
+        val metaScores: Map<String, Int> = mapOf(),
 
         val totalPower: Int = 0,
         val cardDrawCount: Int? = null,
